# project/Services/Admin/adminSwagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for our project."
  },
  "paths": {
    "/luca-app/admin/altaRuta": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "GET /luca-app/admin/altaRuta",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      }
    },
    "/luca-app/admin/addAnnounce": {
      "post": {
        "tags": [
          "API"
        ],
        "summary": "Add Announce",
        "description": "Adds a new announcement to the system.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      }
    },
    "/luca-app/admin/users": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Get Users",
        "description": "Retrieves a list of all users.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      }
    },
    "/luca-app/admin/announces": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Get Announces",
        "description": "Retrieves all announcements.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      },
      "put": {
        "tags": [
          "API"
        ],
        "summary": "Update Announces",
        "description": "Updates an existing announcement.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "API"
        ],
        "summary": "Delete Announces",
        "description": "Deletes an announcement.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      }
    },
    "/luca-app/admin/get-news/:actor": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Get News for Actor",
        "description": "Fetches news related to a specific actor.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      }
    },
    "/luca-app/admin/RSS/:actor": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Get RSS Feed for Actor",
        "description": "Generates an RSS feed with news related to a specific actor.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      }
    },
    "/luca-app/admin/logout": {
      "post": {
        "tags": [
          "API"
        ],
        "summary": "Logout Admin",
        "description": "Logs out an admin user and clears the session.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      }
    },
    "/luca-app/admin/nominatedActors": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Get Nominated Actors",
        "description": "Retrieves a list of actors who have been nominated for awards.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      }
    },
    "/luca-app/admin/user": {
      "delete": {
        "tags": [
          "API"
        ],
        "summary": "Delete User",
        "description": "Deletes a user from the system.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {}
      },
      "put": {
        "tags": [
          "API"
        ],
        "summary": "Update User",
        "description": "Updates the details of an existing user.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "API"
        ],
        "summary": "Add User",
        "description": "Adds a new user to the system.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      }
    },
    "/luca-app/admin/import-csv": {
      "post": {
        "tags": [
          "API"
        ],
        "summary": "Import CSV",
        "description": "Imports user data from a CSV file.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      }
    }
  }
}